requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
pymysql==1.1.0
Werkzeug==3.0.1
